+++ src/Modules/Math/ConvertScalarToMatrix.h
@@ -46,7 +46,7 @@ namespace Math {
     virtual void setStateDefaults() {};
     static const Dataflow::Networks::ModuleLookupInfo staticInfo_;
 
-    INPUT_PORT(0, Input, Double);
+    INPUT_PORT(0, Input, Int32);
     OUTPUT_PORT(0, Output, Matrix);
   };
 }}}
